Unit test for DAGMC cases
